requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
python-dotenv==0.19.2
